Improved Enemy Attacks
Enemies now attack the player when not in the player's POV. Wanderer will do a basic sneak attack. The Speedster is the same but moves much faster and also drains stamina when landing a successful attack. The Phaser will try to teleport out of the player's POV so it can continue to attack.